requirements.txt
Django==1.5.1
South==0.7.6
argparse==1.2.1
dj-database-url==0.2.1
django-tastypie==0.9.15
gunicorn==0.17.4
mimeparse==0.1.3
psycopg2==2.5
python-dateutil==2.1
six==1.3.0
wsgiref==0.1.2
